refactored eventmodal, now modal is a generic component
